refactor client dashboard and api routes
