+++ app/api/clients/[id]/projects/route.ts
@@ -1,50 +1,27 @@
-import { getClientProjects } from '@/lib/apiHandlers/clients';
 import { NextResponse } from 'next/server';
 import type { NextRequest } from 'next/server';
-import { loadUserSession } from '@/lib/loadUserSession';
 import { db } from '@/lib/db';
 import { projects, clientProfiles } from '@/lib/schema';
 import { eq } from 'drizzle-orm';
-import type { SessionClaimsWithRole } from '@/lib/types';
+import { loadUserSession } from '@/lib/loadUserSession';
 
-type RouteContext = { params: Promise<{ id: string }> };
+export const runtime = 'nodejs';
+export const dynamic = 'force-dynamic';
 
-export async function GET(_req: NextRequest, ctx: RouteContext) {
-  const { id } = await ctx.params;
-  const clerkActive = !!process.env.CLERK_SECRET_KEY;
-  let userId: string | undefined;
-  let sessionClaims: SessionClaimsWithRole | undefined;
-  if (clerkActive) {
-    const { auth } = await import('@clerk/nextjs/server');
-    const result = await auth();
-    userId = result.userId;
-    sessionClaims = result.sessionClaims as SessionClaimsWithRole;
-  }
-  const user_role = sessionClaims?.metadata?.user_role;
-  
-  // Check if user is authenticated
-  if (clerkActive && !userId) {
-    return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
-  }
-  
-  // Check if user is admin or the client themselves
-  const isAdmin = user_role === 'admin';
-  const isOwnProfile = userId === id;
-  
-  if (clerkActive && !isAdmin && !isOwnProfile) {
-    return NextResponse.json({ error: 'Forbidden' }, { status: 403 });
-  }
-  
+export async function GET(req: Request, { params }: { params: { id: string } }) {
+  const url = new URL(req.url);
+  const clientId = url.searchParams.get('override') ?? params.id;
+
   try {
-    const projects = await getClientProjects(id);
-    return NextResponse.json({ projects });
+    const data = await db.select().from(projects).where(eq(projects.clientId, clientId));
+    return NextResponse.json({ projects: data });
   } catch (error) {
     console.error('Error fetching client projects:', error);
     return NextResponse.json({ error: 'Failed to fetch client projects' }, { status: 500 });
   }
 }
 
-export async function POST(req: NextRequest, ctx: RouteContext) {
+export async function POST(req: NextRequest, { params }: { params: { id: string } }) {
   const sessionUser = await loadUserSession();
   if (
     !sessionUser ||
@@ -53,8 +30,7 @@ export async function POST(req: NextRequest, ctx: RouteContext) {
     return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
   }
 
-  const { id } = await ctx.params;
-  if (sessionUser.userId !== id) {
+  if (sessionUser.userId !== params.id) {
     return NextResponse.json({ error: 'Forbidden' }, { status: 403 });
   }
